+++ addons/main/CfgWeapons.hpp
@@ -1,3 +1,7 @@
+class asdg_OpticRail1913_long;
+class asdg_OpticRailL1A1;
+class asdg_MuzzleSlot_762;
+
 class CfgWeapons {
     class hgun_P07_F;
     class SMG_02_F;
@@ -54,36 +58,82 @@ class CfgWeapons {
         };
     };
 
+    class Rifle_Long_Base_F: Rifle_Base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class LRR_base_F: Rifle_Long_Base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class srifle_LRR_F: LRR_base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class UK3CB_BAF_L115_Base: srifle_LRR_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {
+            class CowsSlot: asdg_OpticRail1913_long {};
+        };
+    };
+    class UK3CB_BAF_L118A1_Covert: UK3CB_BAF_L115_Base {
+        class WeaponSlotsInfo: WeaponSlotsInfo {
+            class CowsSlot: asdg_OpticRail1913_long {};
+        };
+    };
+
+    class EBR_base_F: Rifle_Long_Base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class srifle_EBR_F: EBR_base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+
     class UK3CB_BAF_L128A1: srifle_EBR_F {
         magazineWell[] += {"CBA_12g_1rnd","CBA_12g_2rnds","CBA_12g_3rnds","CBA_12g_4rnds","CBA_12g_5rnds","CBA_12g_6rnds","CBA_12g_7rnds","CBA_12g_8rnds"};
     };
 
     class UK3CB_BAF_L129A1: srifle_EBR_F {
         magazineWell[] += {"CBA_762x51_SR25","CBA_762x51_AR10"};
+        class WeaponSlotsInfo: WeaponSlotsInfo {
+            class asdg_OpticRail_UK3CB_BAF_L129A1: asdg_OpticRail1913_long {};
+        };
+    };
+
+    class GM6_base_F: Rifle_Long_Base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class srifle_GM6_F: GM6_base_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {};
+    };
+    class UK3CB_BAF_L135A1: srifle_GM6_F {
+        class WeaponSlotsInfo: WeaponSlotsInfo {
+            class asdg_OpticRail_rifles: asdg_OpticRail1913_long {};
+        };
     };
 
     class UK3CB_BAF_L1A1: Rifle_Base_F {
         magazineWell[] += {"CBA_762x51_FAL","CBA_762x51_FAL_L"};
         class WeaponSlotsInfo: WeaponSlotsInfo {
             mass = 95;
+            delete CowsSlot;
+            class CowsSlot: asdg_OpticRailL1A1 {};
+            class asdg_MuzzleSlot_762FAL: asdg_MuzzleSlot_762 {};
         };
     };
     class UK3CB_BAF_L1A1_Wood: UK3CB_BAF_L1A1 {
         class WeaponSlotsInfo: WeaponSlotsInfo {
             mass = 95;
+            class CowsSlot: asdg_OpticRailL1A1 {};
         };
     };
-
     class UK3CB_BAF_L7A2: Rifle_Base_F {
         ACE_Overheating_allowSwapBarrel = 1;
+        magazineReloadTime = 0;
         magazineWell[] += {"M240_762x51","CBA_762x51_LINKS"};
     };
 
     class UK3CB_BAF_Tripod: UK3CB_BAF_Slung_Static_Weapon_Base {
-  		    magazineWell[] += {"CBA_762x51_LINKS","CBA_50BMG_100rnds"};
+        magazineWell[] += {"CBA_762x51_LINKS","CBA_50BMG_100rnds"};
    	};
    	class UK3CB_BAF_L16_Tripod: UK3CB_BAF_Tripod {
-   		   magazineWell[] = {};
+        magazineWell[] = {};
    	};
 
     class UK3CB_BAF_L85A2: arifle_Mk20_plain_F {
@@ -93,7 +143,7 @@ class CfgWeapons {
 
     class UK3CB_BAF_L22: UK3CB_BAF_L85A2 {
         magazineWell[] += {"STANAG_556x45","CBA_556x45_STANAG","CBA_556x45_STANAG_2D","CBA_556x45_STANAG_2D_XL"};
-        class WeaponSlotsInfo {
+        class WeaponSlotsInfo: WeaponSlotsInfo {
             mass = 78;
         };
     };
@@ -107,6 +157,7 @@ class CfgWeapons {
         magazineWell[] += {"STANAG_556x45","CBA_556x45_STANAG","CBA_556x45_STANAG_2D","CBA_556x45_STANAG_2D_XL"};
         class WeaponSlotsInfo {
             mass = 82;
+            class CowsSlot: asdg_OpticRail1913_long {};
         };
     };
     class UK3CB_BAF_L86A2: UK3CB_BAF_L85A2 {
@@ -118,19 +169,22 @@ class CfgWeapons {
     };
 
     class UK3CB_BAF_L91A1: SMG_02_F {
+        displayName = "L91A1";
         magazineWell[] += {"CBA_9x19_MP5"};
         class WeaponSlotsInfo {
             mass = 68.3;
         };
     };
     class UK3CB_BAF_L92A1: UK3CB_BAF_L91A1 {
+        displayName = "L92A1";
         class WeaponSlotsInfo: WeaponSlotsInfo {
             mass = 75;
+            class asdg_MuzzleSlot_UK3CB_BAF_L91A1 {};
         };
     };
 
     class UK3CB_BAF_L131A1: hgun_P07_F {
-  		    magazineWell[] += {"CBA_9x19_Glock_Full"};
+        magazineWell[] += {"CBA_9x19_Glock_Full"};
     };
     class UK3CB_BAF_L9A1: hgun_P07_F {
         magazines[] += {"UK3CB_BAF_9_15Rnd","UK3CB_BAF_9_17Rnd"};